pyproject.toml
[tool.poetry]
name = "kronos"
version = "0.1.0"
description = "Manage events across many calendars"
authors = ["Piotr Gaczkowski <doomhammerng@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.3.2"
pytz = "^2023.3.post1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
